# Fix a major performance bug in 7.0 re: filter compatibility