@pdillinger 2022-03-23
I also tested -format_version=4 for legacy Bloom filter and -use_ribbon_filter for Ribbon filter. Both work as expected with the fix. Also as discussed internally, table properties use new names:

```
$ sst_dump --file=/dev/shm/rocksdb.fixed/000033.sst --show_properties | grep 'filter policy'
  filter policy name: ribbonfilter
```